checking packages
