+++ .gitignore
@@ -11,3 +11,5 @@ src/test/java/com/modus/edeliveryclient/SDBWrapperTest.java
 C:/eclipseProj/edelivery/EDeliveryClient/src/test/resources/standardBusinessDocumentAllXMLtest.xsd
 
 C:/eclipseProj/edelivery/EDeliveryClient/src/test/resources/standardBusinessDocumentAllXMLtest.xsd
+
+C:/eclipseProj/edelivery/EDeliveryClient/src/test/resources/standardBusinessDocumentAllXMLtest.xsd
