+++ src/main/java/com/modus/edeliveryclient/jaxb/assertion/AdviceType.java
@@ -6,7 +6,7 @@
 //
 
 
-package oasis.names.tc.saml._2_0.assertion;
+package com.modus.edeliveryclient.jaxb.assertion;
 
 import java.util.ArrayList;
 import java.util.List;
