+++ src/main/java/com/modus/edeliveryclient/jaxb/assertion/OneTimeUseType.java
@@ -6,24 +6,23 @@
 //
 
 
-package oasis.names.tc.saml._2_0.assertion;
+package com.modus.edeliveryclient.jaxb.assertion;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlSeeAlso;
 import javax.xml.bind.annotation.XmlType;
 
 
 /**
- * <p>Java class for StatementAbstractType complex type.
+ * <p>Java class for OneTimeUseType complex type.
  * 
  * <p>The following schema fragment specifies the expected content contained within this class.
  * 
  * <pre>
- * &lt;complexType name="StatementAbstractType">
+ * &lt;complexType name="OneTimeUseType">
  *   &lt;complexContent>
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
- *     &lt;/restriction>
+ *     &lt;extension base="{urn:oasis:names:tc:SAML:2.0:assertion}ConditionAbstractType">
+ *     &lt;/extension>
  *   &lt;/complexContent>
  * &lt;/complexType>
  * </pre>
@@ -31,13 +30,10 @@ import javax.xml.bind.annotation.XmlType;
  * 
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "StatementAbstractType")
-@XmlSeeAlso({
-    AuthzDecisionStatementType.class,
-    AttributeStatementType.class,
-    AuthnStatementType.class
-})
-public abstract class StatementAbstractType {
+@XmlType(name = "OneTimeUseType")
+public class OneTimeUseType
+    extends ConditionAbstractType
+{
 
 
 }
