+++ src/main/java/com/modus/edeliveryclient/jaxb/assertion/StatementAbstractType.java
@@ -6,25 +6,23 @@
 //
 
 
-package oasis.names.tc.saml._2_0.assertion;
+package com.modus.edeliveryclient.jaxb.assertion;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlSeeAlso;
 import javax.xml.bind.annotation.XmlType;
 
 
 /**
- * <p>Java class for KeyInfoConfirmationDataType complex type.
+ * <p>Java class for StatementAbstractType complex type.
  * 
  * <p>The following schema fragment specifies the expected content contained within this class.
  * 
  * <pre>
- * &lt;complexType name="KeyInfoConfirmationDataType">
+ * &lt;complexType name="StatementAbstractType">
  *   &lt;complexContent>
- *     &lt;restriction base="{urn:oasis:names:tc:SAML:2.0:assertion}SubjectConfirmationDataType">
- *       &lt;sequence>
- *         &lt;element ref="{http://www.w3.org/2000/09/xmldsig#}KeyInfo" maxOccurs="unbounded"/>
- *       &lt;/sequence>
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *     &lt;/restriction>
  *   &lt;/complexContent>
  * &lt;/complexType>
@@ -33,10 +31,13 @@ import javax.xml.bind.annotation.XmlType;
  * 
  */
 @XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "KeyInfoConfirmationDataType")
-public class KeyInfoConfirmationDataType
-    extends SubjectConfirmationDataType
-{
+@XmlType(name = "StatementAbstractType")
+@XmlSeeAlso({
+    AuthzDecisionStatementType.class,
+    AttributeStatementType.class,
+    AuthnStatementType.class
+})
+public abstract class StatementAbstractType {
 
 
 }
